Rename colorPalette to colorPaletteTypes
