+++ src/types/colorPaletteTypes.ts
